requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.2
django-ninja==1.4.1
django-phonenumber-field==8.1.0
phonenumbers==9.0.4
pydantic==2.11.3
pydantic_core==2.33.1
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2